+++ test/issues.js
@@ -12,4 +12,5 @@ describe('Issues.', function () {
   require('./issues/issue-54.js');
   require('./issues/issue-64.js');
   require('./issues/issue-parse-function-security.js');
+  require('./issues/issue-skip-invalid.js');
 });